Changes related with order

+++ client/cards/src/hooks/useForm.jsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 
 import { deleteAllDataPurchaseAfterSendOrder } from '../functions/localStorageFunction/deleteAllDataPurchaseAfterSendOrder';
+import { validateOrder } from '../functions/validateOrder';
 
 export const useForm = (initialValues, onSubmitHandler) => {
     const [values, setValues] = useState(initialValues);
@@ -16,8 +17,11 @@ export const useForm = (initialValues, onSubmitHandler) => {
         onSubmitHandler(values);
 
         setValues(initialValues);
-        
-        deleteAllDataPurchaseAfterSendOrder();
+
+        if (validateOrder(values)) {
+            deleteAllDataPurchaseAfterSendOrder();
+        }
+
     };
 
     const changeValues = (newValues) => {
